requirements.txt
natsort>=7.0.1
librosa>=0.7.2
dill>=0.3.1.1
python_speech_features>=0.6
tqdm>=4.45.0
click>=7.1.1
pandas>=1.0.3
matplotlib>=3.2.1
tensorflow
Keras>=2.3.1
numpy>=1.18.2
numba>=0.48
grpcio
protobuf
